Prepare release 0.23.0

+++ powerauth-java-crypto/src/main/java/io/getlime/security/powerauth/crypto/lib/util/CRC16.java
@@ -97,7 +97,7 @@ public class CRC16 implements Checksum {
     @Override
     public void update(byte[] bytes, int offset, int length) {
         for (int i = offset; i < offset + length; i++) {
-            update((int) bytes[i]);
+            update(bytes[i]);
         }
     }
